move class to correct package
